move ngroup, ffill, pct_change
