+++ python/cudf/cudf/tests/test_groupby.py
@@ -13,7 +13,6 @@ import pytest
 
 import cudf
 from cudf import DataFrame, Series
-from cudf.api.extensions import no_default
 from cudf.core._compat import (
     PANDAS_CURRENT_SUPPORTED_VERSION,
     PANDAS_GE_220,
@@ -908,360 +907,6 @@ def test_groupby_apply_series_args(func, args):
     assert_groupby_results_equal(expect, got)
 
 
-@pytest.mark.parametrize("label", [None, "left", "right"])
-@pytest.mark.parametrize("closed", [None, "left", "right"])
-def test_groupby_freq_week(label, closed):
-    pdf = pd.DataFrame(
-        {
-            "Publish date": [
-                pd.Timestamp("2000-01-03"),
-                pd.Timestamp("2000-01-01"),
-                pd.Timestamp("2000-01-09"),
-                pd.Timestamp("2000-01-02"),
-                pd.Timestamp("2000-01-07"),
-                pd.Timestamp("2000-01-16"),
-            ],
-            "ID": [0, 1, 2, 3, 4, 5],
-            "Price": [10, 20, 30, 40, 50, 60],
-        }
-    )
-    gdf = cudf.from_pandas(pdf)
-    expect = pdf.groupby(
-        pd.Grouper(key="Publish date", freq="1W", label=label, closed=closed)
-    ).mean()
-    got = gdf.groupby(
-        cudf.Grouper(key="Publish date", freq="1W", label=label, closed=closed)
-    ).mean()
-    assert_eq(
-        expect,
-        got,
-        check_like=True,
-        check_dtype=False,
-        check_index_type=False,
-    )
-
-
-@pytest.mark.parametrize("label", [None, "left", "right"])
-@pytest.mark.parametrize("closed", [None, "left", "right"])
-def test_groupby_freq_day(label, closed):
-    pdf = pd.DataFrame(
-        {
-            "Publish date": [
-                pd.Timestamp("2000-01-03"),
-                pd.Timestamp("2000-01-01"),
-                pd.Timestamp("2000-01-09"),
-                pd.Timestamp("2000-01-02"),
-                pd.Timestamp("2000-01-07"),
-                pd.Timestamp("2000-01-16"),
-            ],
-            "ID": [0, 1, 2, 3, 4, 5],
-            "Price": [10, 20, 30, 40, 50, 60],
-        }
-    )
-    gdf = cudf.from_pandas(pdf)
-    expect = pdf.groupby(
-        pd.Grouper(key="Publish date", freq="3D", label=label, closed=closed)
-    ).mean()
-    got = gdf.groupby(
-        cudf.Grouper(key="Publish date", freq="3D", label=label, closed=closed)
-    ).mean()
-    assert_eq(
-        expect,
-        got,
-        check_like=True,
-        check_dtype=False,
-        check_index_type=False,
-    )
-
-
-@pytest.mark.parametrize("label", [None, "left", "right"])
-@pytest.mark.parametrize("closed", [None, "left", "right"])
-def test_groupby_freq_min(label, closed):
-    pdf = pd.DataFrame(
-        {
-            "Publish date": [
-                pd.Timestamp("2000-01-01 12:01:00"),
-                pd.Timestamp("2000-01-01 12:05:00"),
-                pd.Timestamp("2000-01-01 15:30:00"),
-                pd.Timestamp("2000-01-02 00:00:00"),
-                pd.Timestamp("2000-01-01 23:47:00"),
-                pd.Timestamp("2000-01-02 00:05:00"),
-            ],
-            "ID": [0, 1, 2, 3, 4, 5],
-            "Price": [10, 20, 30, 40, 50, 60],
-        }
-    )
-    gdf = cudf.from_pandas(pdf)
-    expect = pdf.groupby(
-        pd.Grouper(key="Publish date", freq="1h", label=label, closed=closed)
-    ).mean()
-    got = gdf.groupby(
-        cudf.Grouper(key="Publish date", freq="1h", label=label, closed=closed)
-    ).mean()
-    assert_eq(
-        expect,
-        got,
-        check_like=True,
-        check_dtype=False,
-        check_index_type=False,
-    )
-
-
-@pytest.mark.parametrize("label", [None, "left", "right"])
-@pytest.mark.parametrize("closed", [None, "left", "right"])
-def test_groupby_freq_s(label, closed):
-    pdf = pd.DataFrame(
-        {
-            "Publish date": [
-                pd.Timestamp("2000-01-01 00:00:02"),
-                pd.Timestamp("2000-01-01 00:00:07"),
-                pd.Timestamp("2000-01-01 00:00:02"),
-                pd.Timestamp("2000-01-02 00:00:15"),
-                pd.Timestamp("2000-01-01 00:00:05"),
-                pd.Timestamp("2000-01-02 00:00:09"),
-            ],
-            "ID": [0, 1, 2, 3, 4, 5],
-            "Price": [10, 20, 30, 40, 50, 60],
-        }
-    )
-    gdf = cudf.from_pandas(pdf)
-    expect = pdf.groupby(
-        pd.Grouper(key="Publish date", freq="3s", label=label, closed=closed)
-    ).mean()
-    got = gdf.groupby(
-        cudf.Grouper(key="Publish date", freq="3s", label=label, closed=closed)
-    ).mean()
-    assert_eq(
-        expect,
-        got,
-        check_like=True,
-        check_dtype=False,
-        check_index_type=False,
-    )
-
-
-@pytest.mark.parametrize(
-    "pdf, group, name, obj",
-    [
-        (
-            pd.DataFrame({"X": ["A", "B", "A", "B"], "Y": [1, 4, 3, 2]}),
-            "X",
-            "A",
-            None,
-        ),
-        (
-            pd.DataFrame({"X": ["A", "B", "A", "B"], "Y": [1, 4, 3, 2]}),
-            "X",
-            "B",
-            None,
-        ),
-        (
-            pd.DataFrame({"X": ["A", "B", "A", "B"], "Y": [1, 4, 3, 2]}),
-            "X",
-            "A",
-            pd.DataFrame({"a": [1, 2, 4, 5, 10, 11]}),
-        ),
-        (
-            pd.DataFrame({"X": ["A", "B", "A", "B"], "Y": [1, 4, 3, 2]}),
-            "Y",
-            1,
-            pd.DataFrame({"a": [1, 2, 4, 5, 10, 11]}),
-        ),
-        (
-            pd.DataFrame({"X": ["A", "B", "A", "B"], "Y": [1, 4, 3, 2]}),
-            "Y",
-            3,
-            pd.DataFrame({"a": [1, 2, 0, 11]}),
-        ),
-    ],
-)
-@pytest.mark.skipif(
-    PANDAS_VERSION < PANDAS_CURRENT_SUPPORTED_VERSION,
-    reason="Warnings only given on newer versions.",
-)
-def test_groupby_get_group(pdf, group, name, obj):
-    gdf = cudf.from_pandas(pdf)
-
-    if isinstance(obj, pd.DataFrame):
-        gobj = cudf.from_pandas(obj)
-    else:
-        gobj = obj
-
-    pgb = pdf.groupby(group)
-    ggb = gdf.groupby(group)
-    with expect_warning_if(obj is not None):
-        expected = pgb.get_group(name=name, obj=obj)
-    with expect_warning_if(obj is not None):
-        actual = ggb.get_group(name=name, obj=gobj)
-
-    assert_groupby_results_equal(expected, actual)
-
-    expected = pdf.iloc[pgb.indices.get(name)]
-    actual = gdf.iloc[ggb.indices.get(name)]
-
-    assert_eq(expected, actual)
-
-
-@pytest.mark.parametrize(
-    "by",
-    [
-        "a",
-        ["a", "b"],
-        pd.Series([2, 1, 1, 2, 2]),
-        pd.Series(["b", "a", "a", "b", "b"]),
-    ],
-)
-@pytest.mark.parametrize("agg", ["sum", "mean", lambda df: df.mean()])
-def test_groupby_transform_aggregation(by, agg):
-    gdf = cudf.DataFrame(
-        {"a": [2, 2, 1, 2, 1], "b": [1, 1, 1, 2, 2], "c": [1, 2, 3, 4, 5]}
-    )
-    pdf = gdf.to_pandas()
-
-    expected = pdf.groupby(by).transform(agg)
-    actual = gdf.groupby(by).transform(agg)
-
-    assert_groupby_results_equal(expected, actual)
-
-
-def test_groupby_select_then_ffill():
-    pdf = pd.DataFrame(
-        {
-            "a": [1, 1, 1, 2, 2],
-            "b": [1, None, None, 2, None],
-            "c": [3, None, None, 4, None],
-        }
-    )
-    gdf = cudf.from_pandas(pdf)
-
-    expected = pdf.groupby("a")["c"].ffill()
-    actual = gdf.groupby("a")["c"].ffill()
-
-    assert_groupby_results_equal(expected, actual)
-
-
-def test_groupby_select_then_shift():
-    pdf = pd.DataFrame(
-        {"a": [1, 1, 1, 2, 2], "b": [1, 2, 3, 4, 5], "c": [3, 4, 5, 6, 7]}
-    )
-    gdf = cudf.from_pandas(pdf)
-
-    expected = pdf.groupby("a")["c"].shift(1)
-    actual = gdf.groupby("a")["c"].shift(1)
-
-    assert_groupby_results_equal(expected, actual)
-
-
-def test_groupby_select_then_diff():
-    pdf = pd.DataFrame(
-        {"a": [1, 1, 1, 2, 2], "b": [1, 2, 3, 4, 5], "c": [3, 4, 5, 6, 7]}
-    )
-    gdf = cudf.from_pandas(pdf)
-
-    expected = pdf.groupby("a")["c"].diff(1)
-    actual = gdf.groupby("a")["c"].diff(1)
-
-    assert_groupby_results_equal(expected, actual)
-
-
-# TODO: Add a test including datetime64[ms] column in input data
-
-
-@pytest.mark.parametrize("by", ["a", ["a", "b"], pd.Series([1, 2, 1, 3])])
-def test_groupby_transform_maintain_index(by):
-    # test that we maintain the index after a groupby transform
-    gdf = cudf.DataFrame(
-        {"a": [1, 1, 1, 2], "b": [1, 2, 1, 2]}, index=[3, 2, 1, 0]
-    )
-    pdf = gdf.to_pandas()
-    assert_groupby_results_equal(
-        pdf.groupby(by).transform("max"), gdf.groupby(by).transform("max")
-    )
-
-
-@pytest.mark.skipif(
-    PANDAS_VERSION < PANDAS_CURRENT_SUPPORTED_VERSION,
-    reason="warning not present in older pandas versions",
-)
-@pytest.mark.parametrize(
-    "data, gkey",
-    [
-        (
-            {
-                "id": ["a", "a", "a", "b", "b", "b", "c", "c", "c"],
-                "val1": [5, 4, 6, 4, 8, 7, 4, 5, 2],
-                "val2": [4, 5, 6, 1, 2, 9, 8, 5, 1],
-                "val3": [4, 5, 6, 1, 2, 9, 8, 5, 1],
-            },
-            ["id"],
-        ),
-        (
-            {
-                "id": [0, 0, 0, 0, 1, 1, 1],
-                "a": [1, 3, 4, 2.0, -3.0, 9.0, 10.0],
-                "b": [10.0, 23, -4.0, 2, -3.0, None, 19.0],
-            },
-            ["id", "a"],
-        ),
-        (
-            {
-                "id": ["a", "a", "b", "b", "c", "c"],
-                "val1": [None, None, None, None, None, None],
-            },
-            ["id"],
-        ),
-    ],
-)
-@pytest.mark.parametrize("periods", [-5, -2, 0, 2, 5])
-@pytest.mark.parametrize("fill_method", ["ffill", "bfill", no_default, None])
-def test_groupby_pct_change(data, gkey, periods, fill_method):
-    gdf = cudf.DataFrame(data)
-    pdf = gdf.to_pandas()
-
-    with expect_warning_if(fill_method not in (no_default, None)):
-        actual = gdf.groupby(gkey).pct_change(
-            periods=periods, fill_method=fill_method
-        )
-    with expect_warning_if(
-        (
-            fill_method not in (no_default, None)
-            or (fill_method is not None and pdf.isna().any().any())
-        )
-    ):
-        expected = pdf.groupby(gkey).pct_change(
-            periods=periods, fill_method=fill_method
-        )
-
-    assert_eq(expected, actual)
-
-
-@pytest.mark.parametrize("periods", [-5, 5])
-def test_groupby_pct_change_multiindex_dataframe(periods):
-    gdf = cudf.DataFrame(
-        {
-            "a": [1, 1, 2, 2],
-            "b": [1, 1, 2, 3],
-            "c": [2, 3, 4, 5],
-            "d": [6, 8, 9, 1],
-        }
-    ).set_index(["a", "b"])
-
-    actual = gdf.groupby(level=["a", "b"]).pct_change(periods)
-    expected = gdf.to_pandas().groupby(level=["a", "b"]).pct_change(periods)
-
-    assert_eq(expected, actual)
-
-
-def test_groupby_pct_change_empty_columns():
-    gdf = cudf.DataFrame(columns=["id", "val1", "val2"])
-    pdf = gdf.to_pandas()
-
-    actual = gdf.groupby("id").pct_change()
-    expected = pdf.groupby("id").pct_change()
-
-    assert_eq(expected, actual)
-
-
 @pytest.mark.parametrize("group_keys", [None, True, False])
 @pytest.mark.parametrize("by", ["A", ["A", "B"]])
 def test_groupby_group_keys(group_keys, by):
@@ -1280,54 +925,6 @@ def test_groupby_group_keys(group_keys, by):
     actual = g_group[["B", "C"]].apply(lambda x: x / x.sum())
     expected = p_group[["B", "C"]].apply(lambda x: x / x.sum())
     assert_eq(actual, expected)
-
-
-@pytest.mark.parametrize(
-    "by",
-    [
-        lambda: "a",
-        lambda: "b",
-        lambda: ["a", "b"],
-        lambda: "c",
-        lambda: pd.Series([1, 2, 1, 2, 1, 2]),
-        lambda: pd.Series(["x", "y", "y", "x", "z", "x"]),
-    ],
-)
-@pytest.mark.parametrize("ascending", [True, False])
-def test_groupby_ngroup(by, ascending):
-    df_ngroup = cudf.DataFrame(
-        {
-            "a": [2, 2, 1, 1, 2, 3],
-            "b": [1, 2, 1, 2, 1, 2],
-            "c": ["a", "a", "b", "c", "d", "c"],
-        },
-        index=[1, 3, 5, 7, 4, 2],
-    )
-    df_ngroup.index.name = "foo"
-    by = by()
-    expected = df_ngroup.to_pandas().groupby(by).ngroup(ascending=ascending)
-    actual = df_ngroup.groupby(by).ngroup(ascending=ascending)
-    assert_eq(expected, actual, check_dtype=False)
-
-
-@pytest.mark.skipif(
-    PANDAS_VERSION < PANDAS_CURRENT_SUPPORTED_VERSION,
-    reason="warning not present in older pandas versions",
-)
-@pytest.mark.parametrize(
-    "groups", ["a", "b", "c", ["a", "c"], ["a", "b", "c"]]
-)
-def test_groupby_dtypes(groups):
-    df = cudf.DataFrame(
-        {"a": [1, 2, 3, 3], "b": ["x", "y", "z", "a"], "c": [10, 11, 12, 12]}
-    )
-    pdf = df.to_pandas()
-    with pytest.warns(FutureWarning):
-        expected = pdf.groupby(groups).dtypes
-    with pytest.warns(FutureWarning):
-        actual = df.groupby(groups).dtypes
-
-    assert_eq(expected, actual)
 
 
 @pytest.mark.parametrize("index_names", ["a", "b", "c", ["b", "c"]])
